refactor(ci): Move CIJobRequest and rename it CIStageDefinition
This definition was originally only used one specific task, but
it really represents the full stage definition of a CIStage, so
rename it and move it to a more appropriate location.
