+++ orca-igor/src/test/groovy/com/netflix/spinnaker/orca/igor/tasks/RetryableIgorTaskSpec.groovy
@@ -18,6 +18,7 @@ package com.netflix.spinnaker.orca.igor.tasks
 
 import com.netflix.spinnaker.orca.ExecutionStatus
 import com.netflix.spinnaker.orca.TaskResult
+import com.netflix.spinnaker.orca.igor.model.CIStageDefinition
 import com.netflix.spinnaker.orca.pipeline.model.Stage
 import retrofit.RetrofitError
 import retrofit.client.Response
@@ -25,9 +26,9 @@ import spock.lang.Specification
 import spock.lang.Subject
 
 class RetryableIgorTaskSpec extends Specification {
-  CIJobRequest jobRequest = Stub(CIJobRequest)
+  CIStageDefinition jobRequest = Stub(CIStageDefinition)
   Stage stage = Mock(Stage) {
-    mapTo(CIJobRequest.class) >> jobRequest
+    mapTo(CIStageDefinition.class) >> jobRequest
   }
 
   @Subject
